refactor: split document responsibilities
