+++ internal/persis/file/doc/store.go
@@ -10,7 +10,6 @@ import (
 	"io"
 	"os"
 	"path/filepath"
-	"regexp"
 	"slices"
 	"sort"
 	"strings"
@@ -22,7 +21,6 @@ import (
 	"github.com/dagucloud/dagu/v2/internal/cmn/logger/tag"
 	"github.com/dagucloud/dagu/v2/internal/core/docs"
 	"github.com/dagucloud/dagu/v2/internal/core/exec"
-	"github.com/dagucloud/dagu/v2/internal/persis/file/dag/grep"
 	"github.com/goccy/go-yaml"
 )
 
@@ -1305,524 +1303,6 @@ func (s *Store) renameDirectoryLocked(ctx context.Context, oldID, newID, oldDirP
 	s.cleanEmptyParents(filepath.Dir(oldDirPath))
 	s.rebuildIndexAfterMutation(ctx)
 	return nil
-}
-
-// Search searches all docs for the given query pattern.
-func (s *Store) Search(ctx context.Context, query string) ([]*docs.DocSearchResult, error) {
-	var results []*docs.DocSearchResult
-
-	candidates, err := s.listSearchCandidates(ctx, "")
-	if err != nil {
-		return nil, err
-	}
-	for _, candidate := range candidates {
-		if ctx.Err() != nil {
-			return nil, ctx.Err()
-		}
-		data, _, err := readRegularDocFile(candidate.AbsPath)
-		if err != nil {
-			logger.Warn(ctx, "Failed to read doc while searching", tag.File(candidate.RelPath), tag.Error(err))
-			continue
-		}
-
-		matches, err := grep.Grep(data, query, grep.DefaultGrepOptions)
-		if err != nil {
-			continue // no match or error — skip
-		}
-
-		doc, parseErr := parseDocFile(data, candidate.ID)
-		title := candidate.ID
-		var description string
-		if parseErr == nil {
-			title = doc.Title
-			description = doc.Description
-		}
-
-		results = append(results, &docs.DocSearchResult{
-			ID:          candidate.ID,
-			Title:       title,
-			Description: description,
-			Matches:     matches,
-		})
-	}
-
-	sort.Slice(results, func(i, j int) bool {
-		return results[i].ID < results[j].ID
-	})
-
-	return results, nil
-}
-
-func docSearchPattern(query string) string {
-	return fmt.Sprintf("(?i)%s", regexp.QuoteMeta(query))
-}
-
-type docSearchCursor struct {
-	Version       int      `json:"v"`
-	Query         string   `json:"q"`
-	PathPrefix    string   `json:"prefix,omitempty"`
-	ExcludedRoots []string `json:"exclude,omitempty"`
-	ID            string   `json:"id,omitempty"`
-}
-
-type docMatchCursor struct {
-	Version    int    `json:"v"`
-	Query      string `json:"q"`
-	PathPrefix string `json:"prefix,omitempty"`
-	ID         string `json:"id"`
-	Offset     int    `json:"offset"`
-}
-
-type docSearchCandidate struct {
-	ID      string
-	RelPath string
-	AbsPath string
-}
-
-func (s *Store) listSearchCandidates(ctx context.Context, pathPrefix string) ([]docSearchCandidate, error) {
-	if err := s.ensureFreshIndex(ctx); err != nil {
-		return nil, err
-	}
-
-	s.mu.RLock()
-	candidates := make([]docSearchCandidate, 0, len(s.docs))
-	for _, doc := range s.docs {
-		if !doc.Readable {
-			continue
-		}
-		id, ok := relativeDocID(doc.ID, pathPrefix)
-		if !ok {
-			continue
-		}
-		candidates = append(candidates, docSearchCandidate{
-			ID:      id,
-			RelPath: filepath.ToSlash(id + ".md"),
-			AbsPath: doc.AbsPath,
-		})
-	}
-	s.mu.RUnlock()
-
-	sort.Slice(candidates, func(i, j int) bool {
-		return candidates[i].ID < candidates[j].ID
-	})
-	return candidates, nil
-}
-
-func normalizeExcludedPathRoots(roots []string) []string {
-	if len(roots) == 0 {
-		return nil
-	}
-	normalized := slices.Clone(roots)
-	sort.Strings(normalized)
-	return slices.Compact(normalized)
-}
-
-func decodeDocSearchCursor(raw, query, pathPrefix string, excludedRoots []string) (docSearchCursor, error) {
-	if raw == "" {
-		return docSearchCursor{}, nil
-	}
-	var cursor docSearchCursor
-	if err := exec.DecodeSearchCursor(raw, &cursor); err != nil {
-		return docSearchCursor{}, err
-	}
-	if cursor.Version != docSearchCursorVersion ||
-		cursor.Query != query ||
-		cursor.PathPrefix != pathPrefix ||
-		!slices.Equal(cursor.ExcludedRoots, excludedRoots) {
-		return docSearchCursor{}, exec.ErrInvalidCursor
-	}
-	return cursor, nil
-}
-
-func decodeDocMatchCursor(raw, query, pathPrefix, id string) (docMatchCursor, error) {
-	if raw == "" {
-		return docMatchCursor{ID: id}, nil
-	}
-	var cursor docMatchCursor
-	if err := exec.DecodeSearchCursor(raw, &cursor); err != nil {
-		return docMatchCursor{}, err
-	}
-	if cursor.Version != docSearchCursorVersion || cursor.Query != query || cursor.PathPrefix != pathPrefix || cursor.ID != id || cursor.Offset < 0 {
-		return docMatchCursor{}, exec.ErrInvalidCursor
-	}
-	return cursor, nil
-}
-
-// SearchCursor returns lightweight, cursor-based document search hits.
-func (s *Store) SearchCursor(ctx context.Context, opts docs.SearchDocsOptions) (*exec.CursorResult[docs.DocSearchResult], error) {
-	if opts.Query == "" {
-		return &exec.CursorResult[docs.DocSearchResult]{Items: []docs.DocSearchResult{}}, nil
-	}
-	pathPrefix, err := cleanDocPathPrefix(opts.PathPrefix)
-	if err != nil {
-		return nil, err
-	}
-	excludedRoots := normalizeExcludedPathRoots(opts.ExcludePathRoots)
-
-	cursor, err := decodeDocSearchCursor(opts.Cursor, opts.Query, pathPrefix, excludedRoots)
-	if err != nil {
-		return nil, err
-	}
-
-	limit := max(opts.Limit, 1)
-	matchLimit := max(opts.MatchLimit, 1)
-	results := make([]docs.DocSearchResult, 0, limit)
-	pattern := docSearchPattern(opts.Query)
-	var hasMore bool
-	var nextCursor string
-
-	candidates, err := s.listSearchCandidates(ctx, pathPrefix)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, candidate := range candidates {
-		if ctx.Err() != nil {
-			return nil, ctx.Err()
-		}
-		if cursor.ID != "" && candidate.ID <= cursor.ID {
-			continue
-		}
-		if docPathRootExcluded(candidate.ID, excludedRoots) {
-			continue
-		}
-
-		data, _, err := readRegularDocFile(candidate.AbsPath)
-		if err != nil {
-			logger.Warn(ctx, "Failed to read doc while searching", tag.File(candidate.RelPath), tag.Error(err))
-			continue
-		}
-
-		window, err := grep.GrepWindow(data, pattern, grep.GrepOptions{
-			IsRegexp: true,
-			Before:   grep.DefaultGrepOptions.Before,
-			After:    grep.DefaultGrepOptions.After,
-			Limit:    matchLimit,
-		})
-		if err != nil {
-			if errors.Is(err, grep.ErrNoMatch) {
-				continue
-			}
-			logger.Warn(ctx, "Failed to search doc", tag.File(candidate.RelPath), tag.Error(err))
-			continue
-		}
-
-		if len(results) == limit {
-			hasMore = true
-			nextCursor = exec.EncodeSearchCursor(docSearchCursor{
-				Version:       docSearchCursorVersion,
-				Query:         opts.Query,
-				PathPrefix:    pathPrefix,
-				ExcludedRoots: excludedRoots,
-				ID:            results[len(results)-1].ID,
-			})
-			break
-		}
-
-		doc, parseErr := parseDocFile(data, candidate.ID)
-		title := candidate.ID
-		var description string
-		if parseErr == nil {
-			title = doc.Title
-			description = doc.Description
-		}
-		item := docs.DocSearchResult{
-			ID:             candidate.ID,
-			Title:          title,
-			Description:    description,
-			Matches:        window.Matches,
-			HasMoreMatches: window.HasMore,
-		}
-		if window.HasMore {
-			item.NextMatchesCursor = exec.EncodeSearchCursor(docMatchCursor{
-				Version:    docSearchCursorVersion,
-				Query:      opts.Query,
-				PathPrefix: pathPrefix,
-				ID:         candidate.ID,
-				Offset:     window.NextOffset,
-			})
-		}
-		results = append(results, item)
-	}
-
-	return &exec.CursorResult[docs.DocSearchResult]{
-		Items:      results,
-		HasMore:    hasMore,
-		NextCursor: nextCursor,
-	}, nil
-}
-
-// SearchMatches returns cursor-based snippets for one document.
-func (s *Store) SearchMatches(_ context.Context, id string, opts docs.SearchDocMatchesOptions) (*exec.CursorResult[*exec.Match], error) {
-	if err := docs.ValidateDocID(id); err != nil {
-		return nil, err
-	}
-	if opts.Query == "" {
-		return &exec.CursorResult[*exec.Match]{Items: []*exec.Match{}}, nil
-	}
-	pathPrefix, err := cleanDocPathPrefix(opts.PathPrefix)
-	if err != nil {
-		return nil, err
-	}
-
-	cursor, err := decodeDocMatchCursor(opts.Cursor, opts.Query, pathPrefix, id)
-	if err != nil {
-		return nil, err
-	}
-
-	storedID, err := scopedDocID(pathPrefix, id)
-	if err != nil {
-		return nil, err
-	}
-	path, err := s.docFilePath(storedID)
-	if err != nil {
-		return nil, err
-	}
-	data, _, err := readRegularDocFile(path)
-	if err != nil {
-		if os.IsNotExist(err) {
-			return nil, docs.ErrDocNotFound
-		}
-		if errors.Is(err, docs.ErrDocNotFound) {
-			return nil, docs.ErrDocNotFound
-		}
-		return nil, err
-	}
-
-	window, err := grep.GrepWindow(data, docSearchPattern(opts.Query), grep.GrepOptions{
-		IsRegexp: true,
-		Before:   grep.DefaultGrepOptions.Before,
-		After:    grep.DefaultGrepOptions.After,
-		Offset:   cursor.Offset,
-		Limit:    max(opts.Limit, 1),
-	})
-	if err != nil {
-		if errors.Is(err, grep.ErrNoMatch) {
-			return &exec.CursorResult[*exec.Match]{Items: []*exec.Match{}}, nil
-		}
-		return nil, err
-	}
-
-	result := &exec.CursorResult[*exec.Match]{
-		Items:   window.Matches,
-		HasMore: window.HasMore,
-	}
-	if window.HasMore {
-		result.NextCursor = exec.EncodeSearchCursor(docMatchCursor{
-			Version:    docSearchCursorVersion,
-			Query:      opts.Query,
-			PathPrefix: pathPrefix,
-			ID:         id,
-			Offset:     window.NextOffset,
-		})
-	}
-	return result, nil
-}
-
-// normalizeSortParams returns validated sort field and order with defaults.
-func normalizeSortParams(sortField docs.DocSortField, sortOrder docs.DocSortOrder) (string, string) {
-	sf := string(sortField)
-	switch sf {
-	case "name", "type", "mtime":
-		// valid
-	default:
-		sf = "type"
-	}
-	so := string(sortOrder)
-	switch so {
-	case "asc", "desc":
-		// valid
-	default:
-		so = "asc"
-	}
-	return sf, so
-}
-
-// buildTreeFromIndexLocked builds a tree of DocTreeNode from the cached index.
-// s.mu must be held by the caller.
-func (s *Store) buildTreeFromIndexLocked(pathPrefix, sortField, sortOrder string, excludedRoots []string) []*docs.DocTreeNode {
-	dirNodes := make(map[string]*docs.DocTreeNode)
-	dirEntries := make(map[string]docDirIndexEntry)
-	var topLevel []*docs.DocTreeNode
-
-	var ensureDirNode func(id string) *docs.DocTreeNode
-	ensureDirNode = func(id string) *docs.DocTreeNode {
-		if id == "" {
-			return nil
-		}
-		if node, ok := dirNodes[id]; ok {
-			return node
-		}
-		entry := dirEntries[id]
-		node := &docs.DocTreeNode{
-			ID:       id,
-			Name:     filepath.Base(filepath.FromSlash(id)),
-			Type:     "directory",
-			Children: []*docs.DocTreeNode{},
-			ModTime:  entry.ModTime,
-		}
-		dirNodes[id] = node
-		parentID := parentDocID(id)
-		if parentID == "" {
-			topLevel = append(topLevel, node)
-		} else {
-			parent := ensureDirNode(parentID)
-			parent.Children = append(parent.Children, node)
-		}
-		return node
-	}
-
-	dirIDs := make([]string, 0, len(s.dirs))
-	for id := range s.dirs {
-		if id != "" {
-			dirIDs = append(dirIDs, id)
-		}
-	}
-	sort.Strings(dirIDs)
-	for _, fullID := range dirIDs {
-		id, ok := relativeDocID(fullID, pathPrefix)
-		if !ok || docPathRootExcluded(id, excludedRoots) {
-			continue
-		}
-		if id != "" {
-			dirEntries[id] = s.dirs[fullID]
-			ensureDirNode(id)
-		}
-	}
-
-	docIDs := make([]string, 0, len(s.docs))
-	for id := range s.docs {
-		docIDs = append(docIDs, id)
-	}
-	sort.Strings(docIDs)
-	for _, fullID := range docIDs {
-		doc := s.docs[fullID]
-		id, ok := relativeDocID(fullID, pathPrefix)
-		if !ok || docPathRootExcluded(id, excludedRoots) {
-			continue
-		}
-		node := &docs.DocTreeNode{
-			ID:      id,
-			Name:    filepath.Base(filepath.FromSlash(id)) + ".md",
-			Title:   doc.Title,
-			Type:    "file",
-			ModTime: doc.ModTime,
-		}
-		parentID := parentDocID(id)
-		if parentID == "" {
-			topLevel = append(topLevel, node)
-			continue
-		}
-		parent := ensureDirNode(parentID)
-		parent.Children = append(parent.Children, node)
-	}
-
-	if sortField == "mtime" {
-		propagateModTime(topLevel)
-	}
-	sortTreeNodes(topLevel, sortField, sortOrder)
-	return topLevel
-}
-
-// propagateModTime recursively sets each directory's ModTime to the max of
-// its own ModTime and all descendant ModTimes.
-func propagateModTime(nodes []*docs.DocTreeNode) time.Time {
-	var maxTime time.Time
-	for _, node := range nodes {
-		t := node.ModTime
-		if len(node.Children) > 0 {
-			childMax := propagateModTime(node.Children)
-			if childMax.After(t) {
-				t = childMax
-			}
-			node.ModTime = t
-		}
-		if t.After(maxTime) {
-			maxTime = t
-		}
-	}
-	return maxTime
-}
-
-func compareNodeNames(a, b *docs.DocTreeNode) int {
-	if cmp := strings.Compare(strings.ToLower(a.Name), strings.ToLower(b.Name)); cmp != 0 {
-		return cmp
-	}
-	return strings.Compare(a.ID, b.ID)
-}
-
-func compareNodeModTime(a, b *docs.DocTreeNode) int {
-	switch {
-	case a.ModTime.Before(b.ModTime):
-		return -1
-	case a.ModTime.After(b.ModTime):
-		return 1
-	default:
-		return compareNodeNames(a, b)
-	}
-}
-
-func reverseCompare(cmp int) int {
-	switch {
-	case cmp < 0:
-		return 1
-	case cmp > 0:
-		return -1
-	default:
-		return 0
-	}
-}
-
-func compareTreeNodes(a, b *docs.DocTreeNode, sortField, sortOrder string) int {
-	switch sortField {
-	case "type":
-		var cmp int
-		switch a.Type {
-		case b.Type:
-			cmp = compareNodeNames(a, b)
-		case "directory":
-			cmp = -1
-		default:
-			cmp = 1
-		}
-		if sortOrder == "desc" {
-			return reverseCompare(cmp)
-		}
-		return cmp
-	case "mtime":
-		if a.Type != b.Type {
-			if a.Type == "directory" {
-				return -1
-			}
-			return 1
-		}
-		if a.Type == "directory" {
-			return compareNodeNames(a, b)
-		}
-		cmp := compareNodeModTime(a, b)
-		if sortOrder == "desc" {
-			return reverseCompare(cmp)
-		}
-		return cmp
-	default: // "name"
-		cmp := compareNodeNames(a, b)
-		if sortOrder == "desc" {
-			return reverseCompare(cmp)
-		}
-		return cmp
-	}
-}
-
-// sortTreeNodes sorts nodes recursively according to the given sort field and order.
-func sortTreeNodes(nodes []*docs.DocTreeNode, sortField, sortOrder string) {
-	sort.Slice(nodes, func(i, j int) bool {
-		return compareTreeNodes(nodes[i], nodes[j], sortField, sortOrder) < 0
-	})
-	for _, node := range nodes {
-		if len(node.Children) > 0 {
-			sortTreeNodes(node.Children, sortField, sortOrder)
-		}
-	}
 }
 
 // cleanEmptyParents removes empty parent directories up to baseDir.
