+++ internal/persis/file/doc/store_test.go
@@ -1960,321 +1960,3 @@ func TestDeleteBatchCleansEmptyParents(t *testing.T) {
 	_, err = os.Stat(filepath.Join(store.baseDir, "deep"))
 	assert.True(t, os.IsNotExist(err))
 }
-
-// --- Sort tests ---
-
-func TestListTreeSortNameAsc(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-
-	require.NoError(t, store.Create(ctx, "c-doc", "c"))
-	require.NoError(t, store.Create(ctx, "a-doc", "a"))
-	require.NoError(t, store.Create(ctx, "b-doc", "b"))
-
-	result, err := store.List(ctx, docs.ListDocsOptions{Page: 1, PerPage: 50, Sort: docs.DocSortFieldName, Order: docs.DocSortOrderAsc})
-	require.NoError(t, err)
-	require.Len(t, result.Items, 3)
-	assert.Equal(t, "a-doc", result.Items[0].ID)
-	assert.Equal(t, "b-doc", result.Items[1].ID)
-	assert.Equal(t, "c-doc", result.Items[2].ID)
-}
-
-func TestListTreeSortNameDesc(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-
-	require.NoError(t, store.Create(ctx, "c-doc", "c"))
-	require.NoError(t, store.Create(ctx, "a-doc", "a"))
-	require.NoError(t, store.Create(ctx, "b-doc", "b"))
-
-	result, err := store.List(ctx, docs.ListDocsOptions{Page: 1, PerPage: 50, Sort: docs.DocSortFieldName, Order: docs.DocSortOrderDesc})
-	require.NoError(t, err)
-	require.Len(t, result.Items, 3)
-	assert.Equal(t, "c-doc", result.Items[0].ID)
-	assert.Equal(t, "b-doc", result.Items[1].ID)
-	assert.Equal(t, "a-doc", result.Items[2].ID)
-}
-
-func TestListTreeSortTypeAsc(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-
-	require.NoError(t, store.Create(ctx, "file-b", "b"))
-	require.NoError(t, store.Create(ctx, "dir-a/child", "c"))
-	require.NoError(t, store.Create(ctx, "file-a", "a"))
-
-	result, err := store.List(ctx, docs.ListDocsOptions{Page: 1, PerPage: 50, Sort: docs.DocSortFieldType, Order: docs.DocSortOrderAsc})
-	require.NoError(t, err)
-	require.Len(t, result.Items, 3)
-	// Directories first.
-	assert.Equal(t, "directory", result.Items[0].Type)
-	assert.Equal(t, "dir-a", result.Items[0].Name)
-	// Then files alphabetically.
-	assert.Equal(t, "file", result.Items[1].Type)
-	assert.Equal(t, "file-a.md", result.Items[1].Name)
-	assert.Equal(t, "file", result.Items[2].Type)
-	assert.Equal(t, "file-b.md", result.Items[2].Name)
-}
-
-func TestListTreeSortTypeDesc(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-
-	require.NoError(t, store.Create(ctx, "file-b", "b"))
-	require.NoError(t, store.Create(ctx, "dir-a/child", "c"))
-	require.NoError(t, store.Create(ctx, "file-a", "a"))
-
-	result, err := store.List(ctx, docs.ListDocsOptions{Page: 1, PerPage: 50, Sort: docs.DocSortFieldType, Order: docs.DocSortOrderDesc})
-	require.NoError(t, err)
-	require.Len(t, result.Items, 3)
-	// Files first (desc reverses).
-	assert.Equal(t, "file", result.Items[0].Type)
-	assert.Equal(t, "file", result.Items[1].Type)
-	assert.Equal(t, "directory", result.Items[2].Type)
-}
-
-func TestListTreeSortMtimeDesc(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-
-	require.NoError(t, store.Create(ctx, "old", "old"))
-	require.NoError(t, store.Create(ctx, "mid", "mid"))
-	require.NoError(t, store.Create(ctx, "new", "new"))
-
-	now := time.Now()
-	setModTime(t, filepath.Join(store.baseDir, "old.md"), now.Add(-2*time.Hour))
-	setModTime(t, filepath.Join(store.baseDir, "mid.md"), now.Add(-1*time.Hour))
-	setModTime(t, filepath.Join(store.baseDir, "new.md"), now)
-
-	result, err := store.List(ctx, docs.ListDocsOptions{Page: 1, PerPage: 50, Sort: docs.DocSortFieldMTime, Order: docs.DocSortOrderDesc})
-	require.NoError(t, err)
-	require.Len(t, result.Items, 3)
-	assert.Equal(t, "new", result.Items[0].ID)
-	assert.Equal(t, "mid", result.Items[1].ID)
-	assert.Equal(t, "old", result.Items[2].ID)
-}
-
-func TestListTreeSortMtimeAsc(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-
-	require.NoError(t, store.Create(ctx, "old", "old"))
-	require.NoError(t, store.Create(ctx, "mid", "mid"))
-	require.NoError(t, store.Create(ctx, "new", "new"))
-
-	now := time.Now()
-	setModTime(t, filepath.Join(store.baseDir, "old.md"), now.Add(-2*time.Hour))
-	setModTime(t, filepath.Join(store.baseDir, "mid.md"), now.Add(-1*time.Hour))
-	setModTime(t, filepath.Join(store.baseDir, "new.md"), now)
-
-	result, err := store.List(ctx, docs.ListDocsOptions{Page: 1, PerPage: 50, Sort: docs.DocSortFieldMTime, Order: docs.DocSortOrderAsc})
-	require.NoError(t, err)
-	require.Len(t, result.Items, 3)
-	assert.Equal(t, "old", result.Items[0].ID)
-	assert.Equal(t, "mid", result.Items[1].ID)
-	assert.Equal(t, "new", result.Items[2].ID)
-}
-
-func TestListTreeSortMtimeWithDirectories(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-
-	require.NoError(t, store.Create(ctx, "dir-old/child", "old child"))
-	require.NoError(t, store.Create(ctx, "dir-new/child", "new child"))
-
-	now := time.Now()
-	setModTime(t, filepath.Join(store.baseDir, "dir-old", "child.md"), now.Add(-2*time.Hour))
-	setModTime(t, filepath.Join(store.baseDir, "dir-new", "child.md"), now)
-
-	result, err := store.List(ctx, docs.ListDocsOptions{Page: 1, PerPage: 50, Sort: docs.DocSortFieldMTime, Order: docs.DocSortOrderDesc})
-	require.NoError(t, err)
-	require.Len(t, result.Items, 2)
-	// Directory with newest child first.
-	assert.Equal(t, "dir-new", result.Items[0].ID)
-	assert.Equal(t, "dir-old", result.Items[1].ID)
-	// Verify directory mtime = max of descendants.
-	assert.False(t, result.Items[0].ModTime.Before(result.Items[1].ModTime))
-}
-
-func TestListTreeSortMtimeStable(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-
-	require.NoError(t, store.Create(ctx, "b-doc", "b"))
-	require.NoError(t, store.Create(ctx, "a-doc", "a"))
-
-	// Set same mtime.
-	sameTime := time.Date(2025, 1, 1, 12, 0, 0, 0, time.UTC)
-	setModTime(t, filepath.Join(store.baseDir, "a-doc.md"), sameTime)
-	setModTime(t, filepath.Join(store.baseDir, "b-doc.md"), sameTime)
-
-	result, err := store.List(ctx, docs.ListDocsOptions{Page: 1, PerPage: 50, Sort: docs.DocSortFieldMTime, Order: docs.DocSortOrderAsc})
-	require.NoError(t, err)
-	require.Len(t, result.Items, 2)
-	// Stable: fallback to name.
-	assert.Equal(t, "a-doc", result.Items[0].ID)
-	assert.Equal(t, "b-doc", result.Items[1].ID)
-}
-
-func TestListTreeSortDefaultsToTypeAsc(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-
-	require.NoError(t, store.Create(ctx, "file-a", "a"))
-	require.NoError(t, store.Create(ctx, "dir-a/child", "c"))
-
-	// Empty sort/order → defaults to type/asc.
-	result, err := store.List(ctx, docs.ListDocsOptions{Page: 1, PerPage: 50})
-	require.NoError(t, err)
-	require.Len(t, result.Items, 2)
-	assert.Equal(t, "directory", result.Items[0].Type)
-	assert.Equal(t, "file", result.Items[1].Type)
-}
-
-func TestListFlatSortNameDesc(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-
-	require.NoError(t, store.Create(ctx, "a-doc", "a"))
-	require.NoError(t, store.Create(ctx, "c-doc", "c"))
-	require.NoError(t, store.Create(ctx, "b-doc", "b"))
-
-	result, err := store.ListFlat(ctx, docs.ListDocsOptions{Page: 1, PerPage: 50, Sort: docs.DocSortFieldName, Order: docs.DocSortOrderDesc})
-	require.NoError(t, err)
-	require.Len(t, result.Items, 3)
-	assert.Equal(t, "c-doc", result.Items[0].ID)
-	assert.Equal(t, "b-doc", result.Items[1].ID)
-	assert.Equal(t, "a-doc", result.Items[2].ID)
-}
-
-func TestListFlatSortMtime(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-
-	require.NoError(t, store.Create(ctx, "old-doc", "old"))
-	require.NoError(t, store.Create(ctx, "new-doc", "new"))
-
-	now := time.Now()
-	setModTime(t, filepath.Join(store.baseDir, "old-doc.md"), now.Add(-1*time.Hour))
-	setModTime(t, filepath.Join(store.baseDir, "new-doc.md"), now)
-
-	result, err := store.ListFlat(ctx, docs.ListDocsOptions{Page: 1, PerPage: 50, Sort: docs.DocSortFieldMTime, Order: docs.DocSortOrderDesc})
-	require.NoError(t, err)
-	require.Len(t, result.Items, 2)
-	assert.Equal(t, "new-doc", result.Items[0].ID)
-	assert.Equal(t, "old-doc", result.Items[1].ID)
-}
-
-func TestListTreeSortNestedChildren(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-
-	require.NoError(t, store.Create(ctx, "dir/c-child", "c"))
-	require.NoError(t, store.Create(ctx, "dir/a-child", "a"))
-	require.NoError(t, store.Create(ctx, "dir/b-child", "b"))
-
-	result, err := store.List(ctx, docs.ListDocsOptions{Page: 1, PerPage: 50, Sort: docs.DocSortFieldName, Order: docs.DocSortOrderAsc})
-	require.NoError(t, err)
-	require.Len(t, result.Items, 1)
-	require.Len(t, result.Items[0].Children, 3)
-	// Children sorted by name asc.
-	assert.Equal(t, "dir/a-child", result.Items[0].Children[0].ID)
-	assert.Equal(t, "dir/b-child", result.Items[0].Children[1].ID)
-	assert.Equal(t, "dir/c-child", result.Items[0].Children[2].ID)
-}
-
-func TestListTreeSortMtimeKeepsDirectoriesAlphabetical(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-
-	require.NoError(t, store.Create(ctx, "alpha/child", "alpha"))
-	require.NoError(t, store.Create(ctx, "beta/child", "beta"))
-	require.NoError(t, store.Create(ctx, "old-file", "old"))
-	require.NoError(t, store.Create(ctx, "new-file", "new"))
-
-	now := time.Now()
-	setModTime(t, filepath.Join(store.baseDir, "alpha", "child.md"), now.Add(-4*time.Hour))
-	setModTime(t, filepath.Join(store.baseDir, "beta", "child.md"), now.Add(-1*time.Hour))
-	setModTime(t, filepath.Join(store.baseDir, "old-file.md"), now.Add(-3*time.Hour))
-	setModTime(t, filepath.Join(store.baseDir, "new-file.md"), now)
-
-	result, err := store.List(ctx, docs.ListDocsOptions{
-		Page:    1,
-		PerPage: 50,
-		Sort:    docs.DocSortFieldMTime,
-		Order:   docs.DocSortOrderDesc,
-	})
-	require.NoError(t, err)
-	require.Len(t, result.Items, 4)
-
-	assert.Equal(t, "alpha", result.Items[0].ID)
-	assert.Equal(t, "beta", result.Items[1].ID)
-	assert.Equal(t, "new-file", result.Items[2].ID)
-	assert.Equal(t, "old-file", result.Items[3].ID)
-}
-
-func TestListTreeSortPaginationConsistency(t *testing.T) {
-	store := newTestStore(t)
-	ctx := context.Background()
-
-	require.NoError(t, store.Create(ctx, "c-doc", "c"))
-	require.NoError(t, store.Create(ctx, "a-doc", "a"))
-	require.NoError(t, store.Create(ctx, "b-doc", "b"))
-
-	opts := docs.ListDocsOptions{Page: 1, PerPage: 2, Sort: docs.DocSortFieldName, Order: docs.DocSortOrderDesc}
-	page1, err := store.List(ctx, opts)
-	require.NoError(t, err)
-	require.Len(t, page1.Items, 2)
-
-	opts.Page = 2
-	page2, err := store.List(ctx, opts)
-	require.NoError(t, err)
-	require.Len(t, page2.Items, 1)
-
-	// Combined should be c, b, a (name desc).
-	assert.Equal(t, "c-doc", page1.Items[0].ID)
-	assert.Equal(t, "b-doc", page1.Items[1].ID)
-	assert.Equal(t, "a-doc", page2.Items[0].ID)
-}
-
-func TestPropagateModTime(t *testing.T) {
-	now := time.Now()
-	nodes := []*docs.DocTreeNode{
-		{
-			ID: "dir", Name: "dir", Type: "directory",
-			ModTime: now.Add(-10 * time.Hour),
-			Children: []*docs.DocTreeNode{
-				{ID: "dir/old", Name: "old", Type: "file", ModTime: now.Add(-5 * time.Hour)},
-				{
-					ID: "dir/sub", Name: "sub", Type: "directory",
-					ModTime: now.Add(-8 * time.Hour),
-					Children: []*docs.DocTreeNode{
-						{ID: "dir/sub/newest", Name: "newest", Type: "file", ModTime: now},
-					},
-				},
-			},
-		},
-	}
-
-	maxTime := propagateModTime(nodes)
-	assert.Equal(t, now.Unix(), maxTime.Unix())
-	// Top dir should have the deepest descendant's mtime.
-	assert.Equal(t, now.Unix(), nodes[0].ModTime.Unix())
-	// Sub-dir should have its child's mtime.
-	assert.Equal(t, now.Unix(), nodes[0].Children[1].ModTime.Unix())
-}
-
-func TestPropagateModTimeEmptyDir(t *testing.T) {
-	dirTime := time.Date(2025, 6, 1, 12, 0, 0, 0, time.UTC)
-	nodes := []*docs.DocTreeNode{
-		{
-			ID: "empty", Name: "empty", Type: "directory",
-			ModTime:  dirTime,
-			Children: []*docs.DocTreeNode{},
-		},
-	}
-
-	propagateModTime(nodes)
-	// Empty dir keeps its own mtime.
-	assert.Equal(t, dirTime, nodes[0].ModTime)
-}
