+++ internal/service/frontend/api/v1/docs.go
@@ -9,12 +9,9 @@ import (
 	"fmt"
 	"net/http"
 	"net/url"
-	"sort"
 	"strings"
-	"time"
 
 	"github.com/dagucloud/dagu/v2/api/v1"
-	"github.com/dagucloud/dagu/v2/internal/auth"
 	"github.com/dagucloud/dagu/v2/internal/cmn/logger"
 	"github.com/dagucloud/dagu/v2/internal/cmn/logger/tag"
 	"github.com/dagucloud/dagu/v2/internal/core/docs"
@@ -59,265 +56,6 @@ func (a *API) requireDocManagement() error {
 		return errDocStoreNotAvailable
 	}
 	return nil
-}
-
-func validateDocPath(path string) error {
-	if err := docs.ValidateDocID(path); err != nil {
-		return &Error{
-			Code:       api.ErrorCodeBadRequest,
-			Message:    fmt.Sprintf("invalid doc path: %v", err),
-			HTTPStatus: http.StatusBadRequest,
-		}
-	}
-	return nil
-}
-
-func scopedDocPath(workspaceName, path string) (string, error) {
-	if err := validateDocPath(path); err != nil {
-		return "", err
-	}
-	if workspaceName == "" {
-		return path, nil
-	}
-	scoped := workspaceName + "/" + path
-	if err := validateDocPath(scoped); err != nil {
-		return "", err
-	}
-	return scoped, nil
-}
-
-func visibleDocPath(workspaceName, path string) string {
-	if workspaceName == "" {
-		return path
-	}
-	return strings.TrimPrefix(path, workspaceName+"/")
-}
-
-type docWorkspaceVisibility struct {
-	all     bool
-	allowed map[string]struct{}
-	known   map[string]struct{}
-}
-
-func (a *API) knownDocWorkspaceNames(ctx context.Context, required bool) (map[string]struct{}, error) {
-	if a.workspaceStore == nil {
-		if required {
-			return nil, workspaceStoreUnavailable()
-		}
-		return nil, nil
-	}
-	workspaces, err := a.workspaceStore.List(ctx)
-	if err != nil {
-		if required {
-			return nil, fmt.Errorf("failed to list workspaces: %w", err)
-		}
-		return nil, nil
-	}
-	known := make(map[string]struct{}, len(workspaces))
-	for _, ws := range workspaces {
-		known[ws.Name] = struct{}{}
-	}
-	return known, nil
-}
-
-func (a *API) docWorkspaceVisibility(ctx context.Context) (docWorkspaceVisibility, error) {
-	visibility := docWorkspaceVisibility{all: true}
-	known, err := a.knownDocWorkspaceNames(ctx, a.workspaceStore != nil)
-	if err != nil {
-		return visibility, err
-	}
-	visibility.known = known
-	if a.authService == nil {
-		return visibility, nil
-	}
-	user, ok := auth.UserFromContext(ctx)
-	if !ok {
-		return visibility, errAuthRequired
-	}
-	access := auth.NormalizeWorkspaceAccess(user.WorkspaceAccess)
-	if access.All {
-		return visibility, nil
-	}
-	known, err = a.knownDocWorkspaceNames(ctx, true)
-	if err != nil {
-		return visibility, err
-	}
-	visibility.all = false
-	visibility.allowed = make(map[string]struct{}, len(access.Grants))
-	visibility.known = known
-	for _, grant := range access.Grants {
-		visibility.allowed[grant.Workspace] = struct{}{}
-	}
-	return visibility, nil
-}
-
-func (a *API) noWorkspaceDocVisibility(ctx context.Context) (docWorkspaceVisibility, error) {
-	known, err := a.knownDocWorkspaceNames(ctx, a.workspaceStore != nil)
-	if err != nil {
-		return docWorkspaceVisibility{}, err
-	}
-	return docWorkspaceVisibility{
-		allowed: make(map[string]struct{}),
-		known:   known,
-	}, nil
-}
-
-func (a *API) docWorkspaceVisibilityForSelection(ctx context.Context, selection workspaceSelection) (docWorkspaceVisibility, error) {
-	switch selection.mode {
-	case workspaceSelectionAll:
-		return a.docWorkspaceVisibility(ctx)
-	case workspaceSelectionDefault:
-		return a.noWorkspaceDocVisibility(ctx)
-	case workspaceSelectionNamed:
-		if err := a.requireWorkspaceVisible(ctx, selection.workspace); err != nil {
-			return docWorkspaceVisibility{}, err
-		}
-		return docWorkspaceVisibility{all: true}, nil
-	default:
-		return docWorkspaceVisibility{}, badWorkspaceError("invalid workspace")
-	}
-}
-
-func (a *API) docReadScopeForParams(
-	ctx context.Context,
-	workspaceParam *api.Workspace,
-) (string, docWorkspaceVisibility, error) {
-	selection, err := parseWorkspaceSelection(workspaceParam)
-	if err != nil {
-		return "", docWorkspaceVisibility{}, err
-	}
-	visibility, err := a.docWorkspaceVisibilityForSelection(ctx, selection)
-	if err != nil {
-		return "", docWorkspaceVisibility{}, err
-	}
-	if selection.mode == workspaceSelectionNamed {
-		return selection.workspace, visibility, nil
-	}
-	return "", visibility, nil
-}
-
-func docTargetWorkspaceForParam(workspaceParam *api.Workspace) (string, error) {
-	if workspaceParam == nil {
-		return "", nil
-	}
-	raw := string(*workspaceParam)
-	if raw == "" {
-		return "", badWorkspaceError("workspace must not be empty")
-	}
-	switch raw {
-	case "all":
-		return "", badWorkspaceError("workspace=all cannot target a single document")
-	case "default":
-		return "", nil
-	default:
-		return validateWorkspaceParam(raw)
-	}
-}
-
-func (a *API) docPointReadScopeForParams(
-	ctx context.Context,
-	workspaceParam *api.Workspace,
-) (string, docWorkspaceVisibility, error) {
-	workspaceName, err := docTargetWorkspaceForParam(workspaceParam)
-	if err != nil {
-		return "", docWorkspaceVisibility{}, err
-	}
-	if workspaceName == "" {
-		visibility, err := a.noWorkspaceDocVisibility(ctx)
-		if err != nil {
-			return "", docWorkspaceVisibility{}, err
-		}
-		return "", visibility, nil
-	}
-	if err := a.requireWorkspaceVisible(ctx, workspaceName); err != nil {
-		return "", docWorkspaceVisibility{}, err
-	}
-	return workspaceName, docWorkspaceVisibility{all: true}, nil
-}
-
-func docMutationScopeForParams(workspaceParam *api.Workspace) (string, error) {
-	return docTargetWorkspaceForParam(workspaceParam)
-}
-
-func (a *API) scopedDocMutationPath(ctx context.Context, workspaceName, path string) (string, error) {
-	if workspaceName == "" {
-		known, err := a.knownDocWorkspaceNames(ctx, a.workspaceStore != nil)
-		if err != nil {
-			return "", err
-		}
-		if docWorkspaceNameForPath(path, docWorkspaceVisibility{known: known}, true) != "" {
-			return "", badWorkspaceError("path targets a workspace; set workspace")
-		}
-	}
-	return scopedDocPath(workspaceName, path)
-}
-
-func (v docWorkspaceVisibility) knownWorkspace(name string) bool {
-	if name == "" {
-		return false
-	}
-	if v.known != nil {
-		_, ok := v.known[name]
-		return ok
-	}
-	if v.allowed != nil {
-		_, ok := v.allowed[name]
-		return ok
-	}
-	return false
-}
-
-func docWorkspaceNameForPath(path string, visibility docWorkspaceVisibility, includeWorkspaceRoot bool) string {
-	workspaceName, rest, hasSlash := strings.Cut(path, "/")
-	if workspaceName == "" {
-		return ""
-	}
-	if !hasSlash && !includeWorkspaceRoot {
-		return ""
-	}
-	if hasSlash && rest == "" {
-		return ""
-	}
-	if visibility.knownWorkspace(workspaceName) {
-		return workspaceName
-	}
-	return ""
-}
-
-func docWorkspaceValue(workspaceName, path string, visibility docWorkspaceVisibility, includeWorkspaceRoot bool) *string {
-	if workspaceName != "" {
-		return ptrOf(workspaceName)
-	}
-	return optionalString(docWorkspaceNameForPath(path, visibility, includeWorkspaceRoot))
-}
-
-func (v docWorkspaceVisibility) visible(path string) bool {
-	if v.all {
-		return true
-	}
-	workspaceName, _, _ := strings.Cut(path, "/")
-	if workspaceName == "" {
-		return true
-	}
-	if _, ok := v.known[workspaceName]; !ok {
-		return true
-	}
-	_, ok := v.allowed[workspaceName]
-	return ok
-}
-
-func (v docWorkspaceVisibility) excludedPathRoots() []string {
-	if v.all || len(v.known) == 0 {
-		return nil
-	}
-	roots := make([]string, 0, len(v.known))
-	for name := range v.known {
-		if _, ok := v.allowed[name]; !ok {
-			roots = append(roots, name)
-		}
-	}
-	sort.Strings(roots)
-	return roots
 }
 
 // ListDocs returns documents as tree or flat list.
@@ -826,99 +564,4 @@ func (a *API) GetDocContentData(ctx context.Context, docID string) (any, error) 
 		resp.Workspace = docWorkspaceValue(workspaceName, rawID, visibility, false)
 		return resp, nil
 	})
-}
-
-func toDocResponse(doc *docs.Doc) api.DocResponse {
-	resp := api.DocResponse{
-		Id:          doc.ID,
-		Title:       doc.Title,
-		Description: doc.Description,
-		Content:     doc.Content,
-	}
-	if doc.FilePath != "" {
-		resp.FilePath = &doc.FilePath
-	}
-	if t, err := time.Parse(time.RFC3339, doc.CreatedAt); err == nil {
-		resp.CreatedAt = &t
-	}
-	if t, err := time.Parse(time.RFC3339, doc.UpdatedAt); err == nil {
-		resp.UpdatedAt = &t
-	}
-	return resp
-}
-
-func toDocMetadataResponse(m docs.DocMetadata) api.DocMetadataResponse {
-	resp := api.DocMetadataResponse{
-		Id:          m.ID,
-		Title:       m.Title,
-		Description: m.Description,
-	}
-	if !m.ModTime.IsZero() {
-		t := m.ModTime
-		resp.ModifiedAt = &t
-	}
-	return resp
-}
-
-func toDocTreeResponse(node *docs.DocTreeNode) api.DocTreeNodeResponse {
-	return toDocTreeResponseWithWorkspace(node, "", docWorkspaceVisibility{})
-}
-
-func toDocTreeResponseWithWorkspace(
-	node *docs.DocTreeNode,
-	workspaceName string,
-	visibility docWorkspaceVisibility,
-) api.DocTreeNodeResponse {
-	resp := api.DocTreeNodeResponse{
-		Id:        node.ID,
-		Name:      node.Name,
-		Title:     ptrOf(node.Title),
-		Type:      api.DocTreeNodeResponseType(node.Type),
-		Workspace: docWorkspaceValue(workspaceName, node.ID, visibility, node.Type == "directory"),
-	}
-	if !node.ModTime.IsZero() {
-		t := node.ModTime
-		resp.ModifiedAt = &t
-	}
-	if len(node.Children) > 0 {
-		children := make([]api.DocTreeNodeResponse, 0, len(node.Children))
-		for _, child := range node.Children {
-			children = append(children, toDocTreeResponseWithWorkspace(child, workspaceName, visibility))
-		}
-		resp.Children = &children
-	}
-	return resp
-}
-
-// docSortParams extracts sort field and order from typed request params with defaults.
-func docSortParams(sort *api.ListDocsParamsSort, order *api.ListDocsParamsOrder) (docs.DocSortField, docs.DocSortOrder) {
-	s := docs.DocSortFieldType
-	if sort != nil {
-		s = docs.DocSortField(*sort)
-	}
-	o := docs.DocSortOrderAsc
-	if order != nil {
-		o = docs.DocSortOrder(*order)
-	}
-	return s, o
-}
-
-// docSortParamsFromQuery extracts sort field and order from URL query values
-// with validation. Invalid values fall back to defaults.
-func docSortParamsFromQuery(params url.Values) (docs.DocSortField, docs.DocSortOrder) {
-	s := docs.DocSortField(params.Get("sort"))
-	switch s {
-	case docs.DocSortFieldName, docs.DocSortFieldType, docs.DocSortFieldMTime:
-		// valid
-	default:
-		s = docs.DocSortFieldType
-	}
-	o := docs.DocSortOrder(params.Get("order"))
-	switch o {
-	case docs.DocSortOrderAsc, docs.DocSortOrderDesc:
-		// valid
-	default:
-		o = docs.DocSortOrderAsc
-	}
-	return s, o
 }
